COSBETA-8 Move check your answers page and confirmation page to separate notifications controller
