+++ cosmetics-web/app/controllers/manual_entry_controller.rb
@@ -1,7 +1,7 @@
 class ManualEntryController < ApplicationController
   include Wicked::Wizard
 
-  steps :add_product_name, :add_external_reference, :check_your_answers
+  steps :add_product_name, :add_external_reference
 
   def show
     @notification = Notification.find(params[:notification_id])
@@ -10,13 +10,7 @@ class ManualEntryController < ApplicationController
 
   def update
     @notification = Notification.find(params[:notification_id])
-
-    if step == steps.last
-      @notification.submit_notification!
-    elsif !notification_params.nil?
-      @notification.update(notification_params)
-    end
-
+    @notification.update(notification_params)
     render_wizard @notification
   end
 
@@ -25,13 +19,9 @@ class ManualEntryController < ApplicationController
     redirect_to wizard_path(steps.first, notification_id: @notification.id)
   end
 
-  def confirmation
-    @notification = Notification.find(params[:notification_id])
-  end
-
   def finish_wizard_path
     @notification = Notification.find(params[:notification_id])
-    notification_path(@notification) + '/confirmation'
+    edit_notification_path(@notification)
   end
 
 private
