+++ cosmetics-web/app/views/notifications/confirmation.html.slim
