+++ cosmetics-web/spec/controllers/manual_entry_controller_spec.rb
@@ -30,10 +30,10 @@ RSpec.describe ManualEntryController, type: :controller do
       expect(response).to render_template(:add_product_name)
     end
 
-    it "redirects to the confirmation page on finish" do
+    it "redirects to the check your answers page on finish" do
       notification = Notification.create
       get(:show, params: { notification_id: notification.id, id: 'wicked_finish' })
-      expect(response).to redirect_to(notification_path(notification.id) + '/confirmation')
+      expect(response).to redirect_to(edit_notification_path(notification))
     end
   end
 
@@ -50,22 +50,6 @@ RSpec.describe ManualEntryController, type: :controller do
       post(:update, params: { notification_id: notification.id, id: 'add_product_name',
                     notification: { product_name: 'Super Shampoo' } })
       expect(notification.reload.product_name).to eq('Super Shampoo')
-    end
-
-    it "marks the notification as complete on reaching the final step" do
-      notification = Notification.create
-      notification.state = 'draft_complete'
-      notification.save
-      post(:update, params: { notification_id: notification.id, id: 'check_your_answers' })
-      expect(notification.reload.state).to eq('notification_complete')
-    end
-  end
-
-  describe "GET #confirmation" do
-    it "assigns the correct notification" do
-      notification = Notification.create
-      get(:confirmation, params: { notification_id: notification.id })
-      expect(assigns(:notification)).to eq(notification)
     end
   end
 end
