+++ cosmetics-web/spec/controllers/notifications_controller_spec.rb
@@ -1,0 +1,31 @@
+require 'rails_helper'
+
+RSpec.describe NotificationsController, type: :controller do
+    before do
+        authenticate_user
+    end
+
+    describe "GET /confimation" do
+        it "assigns the correct notification" do
+            notification = Notification.create
+            get(:confirmation, params: { id: notification.id })
+            expect(assigns(:notification)).to eq(notification)
+        end
+
+        it "marks the notification as complete" do
+            notification = Notification.create
+            notification.state = 'draft_complete'
+            notification.save
+            get(:confirmation, params: { id: notification.id })
+            expect(notification.reload.state).to eq('notification_complete')
+        end
+    end
+
+    describe "GET /edit" do
+        it "assigns the correct notification" do
+            notification = Notification.create
+            get(:edit, params: { id: notification.id })
+            expect(assigns(:notification)).to eq(notification)
+        end
+    end
+end
